#33 refactor: ArchiveCategory enum을 Archive 엔티티 내부로 이동
