--- src/main/java/com/boaz/backend/domain/archive/entity/ArchiveCategory.java
@@ -1,6 +1,0 @@
-package com.boaz.backend.domain.archive.entity;
-
-public enum ArchiveCategory {
-    // 프로젝트, 기술블로그, 활동사진
-    PROJECT, ACTIVITY, BLOG
-}
